Refactor Report Adapters:  * Move general task execution related tasks to the common package  * create a reusable SchemaMapperAbstract so different type of database can use it as a base
